refactor(quarkus): refactor quarkus and its example
