+++ examples/quarkus/src/main/resources/application.properties
@@ -15,7 +15,7 @@ quarkus.log.category."org.apache.camel".level = INFO
 #quarkus.hawtio.role = admin
 
 #
-# Quarkus property file security configuration
+# Quarkus' property file security configuration
 # https://quarkus.io/guides/security-properties
 #
 quarkus.security.users.embedded.enabled = true
