+++ platforms/quarkus/runtime/src/main/java/io/hawt/quarkus/HawtioConfig.java
@@ -1,13 +1,13 @@
 package io.hawt.quarkus;
 
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
 import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
 import io.quarkus.runtime.annotations.ConfigPhase;
 import io.quarkus.runtime.annotations.ConfigRoot;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 
 @ConfigRoot(name = "hawtio", phase = ConfigPhase.BUILD_AND_RUN_TIME_FIXED)
 public class HawtioConfig {
